requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
datasets>=2.14.0
huggingface-hub>=0.20.0
Pillow>=10.0.0
qwen-vl-utils>=0.0.1
accelerate>=0.25.0
bitsandbytes>=0.41.0
scipy>=1.10.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
